requirements.txt
accelerate==1.10.1
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.6
cssselect==1.3.0
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
deepframe==0.1.1
deeplake==4.3.3
diffusers==0.35.1
dill==0.4.0
distro==1.9.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.2
fsspec==2025.5.1
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google_search_results==2.4.2
googleapis-common-protos==1.70.0
grandalf==0.8
groq==0.26.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.3
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
humbug==0.3.2
idna==3.10
importlib_metadata==8.7.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jaraco.context==6.0.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-experimental==0.3.4
langchain-google-community==2.0.7
langchain-google-genai==2.1.9
langchain-groq==0.3.2
langchain-huggingface==0.3.0
langchain-mistralai==0.2.11
langchain-ollama==0.3.7
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.3.45
language_data==1.3.0
libdeeplake==0.0.164
lxml==6.0.0
lxml_html_clean==0.4.2
lz4==4.4.4
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.18
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.1.0
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numexpr==2.11.0
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
olefile==0.47
ollama==0.5.3
openai==1.95.0
openapi-schema-pydantic==1.2.4
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
parso==0.8.4
pathos==0.3.4
pexpect==4.9.0
pillow==10.4.0
pox==0.3.6
ppft==1.7.7
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.7.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.11.3
safetensors==0.5.3
selenium==4.15.2
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-alignments==0.9.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
streamlit==1.48.0
sympy==1.14.0
tenacity==8.5.0
thinc==8.3.6
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.2
tqdm==4.66.6
traitlets==5.14.3
transformers==4.49.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unstructured==0.18.13
unstructured-client==0.42.3
uritemplate==4.2.0
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.23.0
